+++ contracts/tokens/SecurityToken.sol
@@ -263,7 +263,7 @@ contract SecurityToken is StandardToken, DetailedERC20, ReentrancyGuard, Registr
     * @param _module address of module to unarchive
     */
     function removeModule(address _module) external onlyOwner {
-        require(modulesToData[_module].isArchived, "Module archived");
+        require(modulesToData[_module].isArchived, "Module not archived");
         require(modulesToData[_module].module != address(0), "Module missing");
         emit ModuleRemoved(modulesToData[_module].moduleTypes, _module, now);
         // Remove from module type list
@@ -352,19 +352,23 @@ contract SecurityToken is StandardToken, DetailedERC20, ReentrancyGuard, Registr
     }
 
     /**
-    * @notice allows owner to approve more POLY to one of the modules
+    * @notice allows owner to increase/decrease POLY approval of one of the modules
     * @param _module module address
-    * @param _budget new budget
+    * @param _change change in allowance
+    * @param _increase true if budget has to be increased, false if decrease
     */
-    function changeModuleBudget(address _module, uint256 _budget) external onlyOwner {
+    function changeModuleBudget(address _module, uint256 _change, bool _increase) external onlyOwner {
         require(modulesToData[_module].module != address(0), "Module missing");
-        uint256 _currentAllowance = IERC20(polyToken).allowance(address(this), _module);
-        if (_budget < _currentAllowance) {
-            require(IERC20(polyToken).decreaseApproval(_module, _currentAllowance.sub(_budget)), "Insufficient balance");
+        uint256 currentAllowance = IERC20(polyToken).allowance(address(this), _module);
+        uint256 newAllowance;
+        if (_increase) {
+            require(IERC20(polyToken).increaseApproval(_module, _change), "increaseApproval fail");
+            newAllowance = currentAllowance.add(_change);
         } else {
-            require(IERC20(polyToken).increaseApproval(_module, _budget.sub(_currentAllowance)), "Insufficient balance");
+            require(IERC20(polyToken).decreaseApproval(_module, _change), "Insufficient allowance");
+            newAllowance = currentAllowance.sub(_change);
         }
-        emit ModuleBudgetChanged(modulesToData[_module].moduleTypes, _module, _currentAllowance, _budget);
+        emit ModuleBudgetChanged(modulesToData[_module].moduleTypes, _module, currentAllowance, newAllowance);
     }
 
     /**
@@ -480,7 +484,7 @@ contract SecurityToken is StandardToken, DetailedERC20, ReentrancyGuard, Registr
      * @return bool success
      */
     function transferWithData(address _to, uint256 _value, bytes _data) public returns (bool success) {
-        require(_updateTransfer(msg.sender, _to, _value, _data), "Transfer Invalid");
+        require(_updateTransfer(msg.sender, _to, _value, _data), "Transfer invalid");
         require(super.transfer(_to, _value));
         return true;
     }
@@ -505,7 +509,7 @@ contract SecurityToken is StandardToken, DetailedERC20, ReentrancyGuard, Registr
      * @return bool success
      */
     function transferFromWithData(address _from, address _to, uint256 _value, bytes _data) public returns(bool) {
-        require(_updateTransfer(_from, _to, _value, _data), "Transfer Invalid");
+        require(_updateTransfer(_from, _to, _value, _data), "Transfer invalid");
         require(super.transferFrom(_from, _to, _value));
         return true;
     }
@@ -528,6 +532,10 @@ contract SecurityToken is StandardToken, DetailedERC20, ReentrancyGuard, Registr
     /**
      * @notice validate transfer with TransferManager module if it exists
      * @dev TransferManager module has a key of 2
+     * @dev _isTransfer boolean flag is the deciding factor for whether the 
+     * state variables gets modified or not within the different modules. i.e isTransfer = true
+     * leads to change in the modules environment otherwise _verifyTransfer() works as a read-only
+     * function (no change in the state). 
      * @param _from sender of transfer
      * @param _to receiver of transfer
      * @param _value value of transfer
@@ -620,7 +628,7 @@ contract SecurityToken is StandardToken, DetailedERC20, ReentrancyGuard, Registr
         bytes _data
         ) public onlyModuleOrOwner(MINT_KEY) isMintingAllowed() returns (bool success) {
         require(_investor != address(0), "Investor is 0");
-        require(_updateTransfer(address(0), _investor, _value, _data), "Transfer Invalid");
+        require(_updateTransfer(address(0), _investor, _value, _data), "Transfer invalid");
         _adjustTotalSupplyCheckpoints();
         totalSupply_ = totalSupply_.add(_value);
         balances[_investor] = balances[_investor].add(_value);
@@ -637,7 +645,7 @@ contract SecurityToken is StandardToken, DetailedERC20, ReentrancyGuard, Registr
      * @return success
      */
     function mintMulti(address[] _investors, uint256[] _values) external returns (bool success) {
-        require(_investors.length == _values.length, "Invalid inputs");
+        require(_investors.length == _values.length, "Incorrect inputs");
         for (uint256 i = 0; i < _investors.length; i++) {
             mint(_investors[i], _values[i]);
         }
@@ -678,7 +686,7 @@ contract SecurityToken is StandardToken, DetailedERC20, ReentrancyGuard, Registr
      * @param _data data to indicate validation
      */
     function burnWithData(uint256 _value, bytes _data) public onlyModule(BURN_KEY) {
-        require(_burn(msg.sender, _value, _data), "Burn Invalid");
+        require(_burn(msg.sender, _value, _data), "Burn invalid");
     }
 
     /**
@@ -690,7 +698,7 @@ contract SecurityToken is StandardToken, DetailedERC20, ReentrancyGuard, Registr
     function burnFromWithData(address _from, uint256 _value, bytes _data) public onlyModule(BURN_KEY) {
         require(_value <= allowed[_from][msg.sender], "Value too high");
         allowed[_from][msg.sender] = allowed[_from][msg.sender].sub(_value);
-        require(_burn(_from, _value, _data), "Burn Invalid");
+        require(_burn(_from, _value, _data), "Burn invalid");
     }
 
     /**